ai/requirements.txt
psycopg[binary]
psycopg2-binary
langchain-mistralai
python-dotenv
fastapi
uvicorn
sqlmodel
sqlalchemy
